Verification Service Skeleton
This adapts the skel code to start adding verification service-specific
config and behavior. Of particular interest compared to other services
is that this does not use a database and therefore does not use AR.
The service is backed by an elasticsearch cluster. I've made an attempt
to update the deployment scripts to reflect this. I've also removed
all active-record-y rspec support/setup.

The existing elasticsearch clusters (staging & prod) will remain.

Change-Id: I6b8b31c37137485d2ef6cf4b526dcb4bd4411ab6
Reviewed-on: https://code.brigade.zone/31735
Reviewed-by: Anuj Das <anuj.das@brigade.com>
Reviewed-by: John Miller <john.miller@brigade.com>
Reviewed-by: Ahmad Wilson <ahmad.wilson@brigade.com>
Reviewed-by: Nora Grossman <nora.grossman@brigade.com>
Tested-by: Bryan Eslinger <bryan.eslinger@brigade.com>