Elasticsearch Query Builder
This adds a class and set of helper modules to build the voter record
elasticsearch query to be used by user-initiated voter verification
requests (the "search" rpc defined in the thrift schema, as opposed to
searches from the contact service, which will be added in a separate
commit.) Initially I planned to keep the exact query structure from
the existing service, but as I dug into the query more, I learned the
`filtered` construct that serves as the basis of the current query was
deprecated in ES 2.x (we run 1.7, current is 6.4). We can achieve the
same results with a bool construct even without upgrading ElasticSearch,
so I ended up fundamentally changing the query structure to enable
this upgrade more easily when we decide to do it. (and have more
thoroughly reviewed the change logs to understand what else may
be required).

The basic idea is that logically indivisable chunks of an ES query
are represented by a method contained in a submodule of Clauses.
Those chunks need to be attached to something to work properly with
the elasticsearch-dsl gem, which is achieved via the `parent_clause`
param passed into each of these methods. Then there is a class
`Queries::VoterRecord::Query` that is essentially a factory class
for `Elasticsearch::DSL::Search` objects, exposing methods for
each of the search types available. In the existing vocabulary of
VV, these are "auto" and "top." The objects resulting from these
methods will be passed to an elasticsearch client instance and
used to actually query the ES cluster.

Change-Id: Id313a7900d93272d0b574db645ce5cb6e46a8636
Reviewed-on: https://code.brigade.zone/31933
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>