Preprocessing for ES Queries
This adds modules to hold logic responsible for
transforming data in ways that are specific to building
a query for elastic search: Preprocessors::Name and
Preprocessors::Address.

Address
The address preprocessor handles translating zip codes to
lat/lng strings suitable for use as an ES GeoPoint. To
continue avoiding the need for a mysql database in this
service (and continuing the approach of the existing verifier),
there is unfortunately a 43k-line CSV file containing the mappings
from zip code -> lat/lng. While I'm not proud of the 43k line add
in the commit, this data changes infrequently and has a memory
footprint of ~2MB. The source file is read in from an initializer
leaving, so no request handlers will block on I/O, though this
obviously adds a little bit of startup cost. I think this cost
is more than worth the lookup speed, since the geocoding will
happen from this in-memory hash map. The address preprocessor
also throws away the +4 part of the zip code, if included.

Name
The name preprocessor removes suffixes e.g. Jr./Sr. from
all name fields, as the voterbase data is not consistent
with which name fields contain the suffix data. The name
fields are used in query constructs (i.e. not filter constructs)
so their contributions to match scores without the suffixes
are largely the same. This preprocessor also splits a two-
word first name into first and middle if the provided middle
name is empty.

Change-Id: Iec450c7e69fb0342ff5e2765e055a6c2a9360572
Reviewed-on: https://code.brigade.zone/31870
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Anuj Das <anuj.das@brigade.com>