Configure ES Client
This adds config settings and an initializer for an ES client to be
used talk to the ES cluster, as well as a class method on the
VoterRecord model making use of the client.

This also locks the version of the elasticsearch gem to 1.1,
which is the recommended version to use against ES 1.7 (which is
what we currently run).

Change-Id: I80bfac5c7ec12bdc020b0ff3d411c3cacaa4d841
Reviewed-on: https://code.brigade.zone/32371
Reviewed-by: Anuj Das <anuj.das@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>