VoterRecord Model
The VoterRecord model represents documents of type "voters"
in the "votizen_verifier" index. It is responsible for issuing
queries against the ES cluster, as well as for transforming the
underlying document data into the appropriate thrift structures.

Queries are issued using class methods .get and .search.
The .search method takes a query object (anything that responds
to `to_hash` and returns a hash representing a valid ES query,
but intended usage is to use a `Queries::VoterRecord::Query`
instance. The .get method takes an id and returns a document
matching the id, or nil if none is found. An initializer
configures an Elasticsearch::Client object and sets it on
the VoterRecord class, along with the index name and document
type to use. Theoretically this can be configured via env var,
though in practice these names are determined by the name used
by the Spark job that craetes the index.

The primary instance method is `#to_thrift` and builds a thrift
object to be used by return structures in the search,
contact_search and get_by_identifier RPCs.

Change-Id: Iab2f618492f2e35a2b518059f7be18644a81fb97
Reviewed-on: https://code.brigade.zone/32412
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>