Search Handler & Search Debug Tool
This wires the search RPC handler to the ES querybuilder
and VoterRecord search in what the monorail currently
calls "Smart search mode." This means we will attempt
an "auto-verify" query but fall back to the "top search"
query if no auto-verify match can be found.

This also adds:
> rake search:test last_name=Smith <...other query args...>
> rake search:explain last_name=Smith <...other query args...>

These two rake tasks are useful for debugging ES queries and
results. The task runs "everything but the thrift," which is to say
it builds an elasticsearch query, issues the query, and then builds
the VoterRecord result wrapper instances for the results.

Change-Id: Ia38b77ab9c4f134aa57c7112a25bb5068d09f2c6
Reviewed-on: https://code.brigade.zone/32434
Reviewed-by: Anuj Das <anuj.das@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>