Clarify Separation: VoterRecord::Search / VoterRecord::Query
To clarify the split between building an elasticsearch query,
the business logic of interpreting the results of an elasticsearch
query, and the representation of an elasticsearch voter record document,
this introduces some vocabulary changes (e.g. max_results vs. size) and
refactors the VoterRecord::Search class to be the only place that knows
about whether results are "auto-verified" or not.

All of that came about while implementing result grouping and filtering.
There is a step of result filtering that happens at the application
layer rather than at the elasticsearch layer in order to implement
a notion of equivalence relations to determine when a result set
is above the number of max_results. Note for the mathematically-
pedantic: this is not a true equivalence relation (as it fails the
transitive property and will depend highly on its representative).
But we do roughly define equivalence classes as records having scores
that are within some neighborhood epsilon (called the equivalence width
in code) of other results. When applying the "max_results" filter
in the final step of a search, we only include results if its
entire equivalence class can be included without going over the
max_results threshold.

Change-Id: I0152695689f81b3e550da9913cbedbad50660560
Reviewed-on: https://code.brigade.zone/32482
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>