Query Tuning
First, this adds the previously tuned min_score values
(from the python verifier) and applies them to the ES
queries.

Then, this set the new email score boost to 8. This
represents the idea that an email match alone will
be enough to lift a result over any other partial
matches and is enough to meet the min_score threshold.

This also adds "email" as an accepted field to the rake
test and explain tasks.

Change-Id: If07a1f0a213c3cbc65ff30e9f71dc00a2b3e1c11
Reviewed-on: https://code.brigade.zone/32554
Tested-by: Leeroy Jenkins <jenkins@brigade.com>
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>