Optional Last Name + Phone Number Search
This removes the requirement that a last name be provided
for a top search, instead making the requirement that at
least one of last_name, email or phone is provided.

When no last_name is provided, a filter is added to match
exactly on email or phone number, and a score boost is given
to get above the top score minimum.

Change-Id: I9e97b1b816986198281f539416aeafc7e2ed070f
Reviewed-on: https://code.brigade.zone/32592
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>