Get (Pseudo)Random Addresses RPC
Used to pad lists of addresses used in an identity verification step,
this adds an RPC to take a state and a seed integer and return a list
of (pseudo)randomly selected addresses in that state, based on voter
records. The seed is typically a profile ID, used to ensure a user
gets a consistent list of addresses back. This is to guard against
a fraudulent user repeating the verification request and being able
to see only real answer(s) stay consistent.

Change-Id: I0b7744ff2a92849a8251562be4108069e6a7fc92
Reviewed-on: https://code.brigade.zone/32611
Reviewed-by: Dave Kincade <dave.kincade@brigade.com>
Reviewed-by: Anuj Das <anuj.das@brigade.com>
Tested-by: Leeroy Jenkins <jenkins@brigade.com>