Remove external thrift definition dependency
The thrift definitions are now directly in this repo. This adds
build scripts to use the thrift built-in ruby generator to
generate ruby classes from the thrift defs. Lastly, the Gemfile
is updated to install these generated thrift_defs as a local gem.
Installing as a local gem was chosen as an alternative to name-
spacing the ruby thrift in e.g. Lib.ThriftDefs and require the
files as part of app bootstrap, mostly to avoid an implicit
coupling of the thrift namespace and output location. As a
local gem, the coupling is in the Gemfile (a more natural place).